backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
firebase-admin==6.6.0
google-cloud-storage==2.18.2
python-multipart==0.0.18
python-dotenv==1.0.1
